renames ntl files to be more systematic
