+++ honeybadgermpc/ntl/ccobject.pxd
@@ -1,5 +1,5 @@
 # Code ot help help python objects to C++ objects
-cdef extern from "ccobject.h":
+cdef extern from "ccobject_impl.h":
     # Print representation of any C++ object
     str ccrepr[T](T x)
