+++ honeybadgermpc/ntl/ccobject_impl.h
